+++ src/failsafe.rs
@@ -1,5 +1,8 @@
 #[cfg(feature = "serde")]
-use serde::{Deserialize, Serialize};
+use {
+    sequential_storage::map::PostcardValue,
+    serde::{Deserialize, Serialize},
+};
 
 /// Configuration of failsafe behavior.<br><br>
 #[derive(Clone, Copy, Debug, PartialEq)]
@@ -48,6 +51,9 @@ impl FailsafeConfig {
     }
 }
 
+#[cfg(feature = "serde")]
+impl PostcardValue<'_> for FailsafeConfig {}
+
 impl Default for FailsafeConfig {
     fn default() -> Self {
         Self::new()
@@ -63,10 +69,7 @@ mod tests {
     fn _is_normal<T: Sized + Send + Sync + Unpin>() {}
     fn is_full<T: Sized + Send + Sync + Unpin + Copy + Clone + Default + PartialEq>() {}
     #[cfg(feature = "serde")]
-    fn is_config<
-        T: Sized + Send + Sync + Unpin + Copy + Clone + Default + PartialEq + Serialize + for<'a> Deserialize<'a>,
-    >() {
-    }
+    fn is_config<T: Serialize + for<'a> Deserialize<'a> + for<'a> PostcardValue<'a>>() {}
 
     #[test]
     fn normal_types() {